chore(wls): Add Windows support
